created an on-page switch statement (users / posts) for the selected http methods and the implementation of those methods

+++ rest-api/api/models/users.go
@@ -4,7 +4,7 @@ import "time"
 
 //User is an ...
 type User struct {
-	ID        uint32    `gorm:"primary_key;auto_increment" json:"id"`
+	ID        uint64    `gorm:"primary_key;auto_increment" json:"id"`
 	UserName  string    `gorm:"type:varchar(150);not null;unique_index" json:"user_name"`
 	Email     string    `gorm:"type:varchar(150);not null;unique_index" json:"email"`
 	Password  string    `gorm:"type:varchar(150);not null" json:"password"`
